+++ inst/tinytest/test_as_track.R
@@ -18,3 +18,12 @@ expect_equivalent(d1$y_, as_track(d1s)$y_)
 expect_equivalent(d1$t_, as_track(d1s)$t_)
 expect_equivalent(d1$burst_, as_track(d1s)$burst_)
 expect_equivalent(get_crs(d1), get_crs(as_track(d1s)))
+
+# As track from data.frame
+dat <- data.frame(
+  x_ = 1:10, y_ = 1:10,
+  t_ = lubridate::ymd("2020-02-01") + days(0:9))
+
+expect_equivalent(make_track(dat, x_, y_), as_track(dat[, c("x_", "y_")]))
+expect_equivalent(make_track(dat, x_, y_, t_), as_track(dat[, c("x_", "y_", "t_")]))
+